+++ kamehouse-desktop/bin/effects/drop_shadow_effect.py
@@ -1,11 +1,14 @@
 from PyQt5 import QtGui
 from PyQt5.QtWidgets import QGraphicsDropShadowEffect
-from PyQt5.QtCore import Qt
+
+from config.kamehouse_desktop_cfg import kamehouseDesktopCfg
 
 class DropShadowEffect(QGraphicsDropShadowEffect):
     def __init__(self, item):
         super().__init__()
-        self.setBlurRadius(10)
-        self.setColor(QtGui.QColor("black"))
-        self.setOffset(4,4)
+        self.setBlurRadius(kamehouseDesktopCfg.getInt('drop_shadow_effect', 'blur_radius'))
+        self.setColor(QtGui.QColor(kamehouseDesktopCfg.get('drop_shadow_effect', 'color')))
+        offsetX = kamehouseDesktopCfg.getInt('drop_shadow_effect', 'offset_x')
+        offsetY = kamehouseDesktopCfg.getInt('drop_shadow_effect', 'offset_y')
+        self.setOffset(offsetX, offsetY)
         item.setGraphicsEffect(self)