+++ kamehouse-desktop/bin/widgets/kamehouse_logo.py
@@ -1,12 +1,5 @@
-from PyQt5.QtWidgets import QLabel
-from PyQt5.QtGui import QPixmap
-from loguru import logger
+from widgets.image import ImageWidget
 
-class KameHouseLogoWidget(QLabel):
+class KameHouseLogoWidget(ImageWidget):
     def __init__(self, window):
-        super().__init__(window)
-        logger.info("Initializing kamehouse logo widget")
-        self.setPixmap(QPixmap('lib/ico/kamehouse.png'))
-        self.setGeometry(1660, 100, 60, 60)
-        self.setScaledContents(True) 
-        
+        super().__init__('kamehouse_logo_widget', window)