+++ kamehouse-desktop/bin/widgets/world_cup_logo.py
@@ -1,12 +1,5 @@
-from PyQt5.QtWidgets import QLabel
-from PyQt5.QtGui import QPixmap
-from loguru import logger
+from widgets.image import ImageWidget
 
-class WorldCupLogoWidget(QLabel):
+class WorldCupLogoWidget(ImageWidget):
     def __init__(self, window):
-        super().__init__(window)
-        logger.info("Initializing world cup logo widget")
-        self.setPixmap(QPixmap('lib/ui/img/sports/world-cup.png'))
-        self.setGeometry(90, 980, 45, 45)
-        self.setScaledContents(True) 
-        
+        super().__init__('world_cup_logo_widget', window)